Utility fn to parse SA tag
